refactor: use kebap-case and make it a rule
