+++ src/app.tsx
@@ -1,6 +1,6 @@
-import Map from './Map/Map';
-import MainControl from './Controls';
-import SettingsPanel from './Controls/settings-panel';
+import Map from './map';
+import MainControl from './controls';
+import SettingsPanel from './controls/settings-panel';
 
 export const App = () => {
   return (
