+++ src/components/settings-footer.tsx
