+++ src/controls/directions/index.tsx
@@ -3,27 +3,27 @@ import { connect } from 'react-redux';
 import type { ThunkDispatch } from 'redux-thunk';
 import { Divider, type ButtonProps } from 'semantic-ui-react';
 
-import Waypoints from './Waypoints';
+import Waypoints from './waypoints';
 
 import { ProfilePicker } from '@/components/profile-picker';
-import { SettingsButton } from '@/components/SettingsButton';
-import { SettingsFooter } from '@/components/SettingsFooter';
+import { SettingsButton } from '@/components/settings-button';
+import { SettingsFooter } from '@/components/settings-footer';
 import { Settings } from './settings';
-import { DateTimePicker } from '@/components/DateTimePicker';
+import { DateTimePicker } from '@/components/date-time-picker';
 
 import {
   doAddWaypoint,
   doRemoveWaypoint,
   makeRequest,
   clearRoutes,
-} from '@/actions/directionsActions';
+} from '@/actions/directions-actions';
 import {
   updateProfile,
   doShowSettings,
   updatePermalink,
   resetSettings,
   doUpdateDateTime,
-} from '@/actions/commonActions';
+} from '@/actions/common-actions';
 import type { RootState } from '@/store';
 import type { Profile } from '@/reducers/common';
 import type { AnyAction } from 'redux';
