+++ src/controls/isochrones/contours-information.tsx
