+++ src/controls/isochrones/settings.tsx
